Add more Lootcrate Distance
I added more values for the loot crate distance since there are currently a lot of people wanting a higher range, and for the sake of the people not breaking their antistasi while unpacking with PBO, we could just add it. I see no harm in adding it since in the end the admin/user can still decide what range to use.